requirements.txt
requests>=2.28
rich>=12.0
python-dotenv>=1.0
